bug: 触发事件描述由前端组装 #12969
