+++ src/frontend/locale/codelib/en-US.json
@@ -218,5 +218,76 @@
     "触发器不匹配": "Trigger Not Match",
     "触发结果": "Trigger Result",
     "codelibUndeployTitle": "{0} connection service has not been fully deployed",
-    "githubUndeployDesc": "To use this feature, please contact the platform administrator to register the GitHub APP"
+    "githubUndeployDesc": "To use this feature, please contact the platform administrator to register the GitHub APP",
+    "authorizeAccount": "Authorize Account",
+    "bkTgitPushEventDesc": "[{branch}] Commit [{commit}] pushed by {user}",
+    "bkTgitIssueCreatedEventDesc": "Issue [{issue}] opened by {user}",
+    "bkTgitIssueUpdatedEventDesc": "Issue [{issue}] updated by {user}",
+    "bkTgitIssueClosedEventDesc": "Issue [{issue}] closed by {user}",
+    "bkTgitIssueReopenedEventDesc": "Issue [{issue}] reopened by {user}",
+    "bkTgitMrCreatedEventDesc": "Merge requests [{mr}] opened by {user}",
+    "bkTgitMrUpdatedEventDesc": "Merge requests [{mr}] edited by {user}",
+    "bkTgitMrClosedEventDesc": "Merge requests [{mr}] closed by {user}",
+    "bkTgitMrReopenedEventDesc": "Merge requests [{mr}] reopened by {user}",
+    "bkTgitMrPushUpdatedEventDesc": "Merge requests [{mr}] push updated by {user}",
+    "bkTgitMrMergedEventDesc": "Merge requests [{mr}] merged by {user}",
+    "bkTgitTagPushEventDesc": "[{source}]Tag [{tag}] pushed by {user}",
+    "bkTgitTagDeleteEventDesc": "[{source}]Tag [{tag}] delete by {user}",
+    "bkTgitNoteEventDesc": "Note [{note}] submitted by {user}",
+    "bkTgitReviewCreatedEventDesc": "Code review [{review}] created by {user}",
+    "bkTgitReviewApprovedEventDesc": "Code review [{review}] approved by {user}",
+    "bkTgitReviewApprovingEventDesc": "Code review [{review}] approving",
+    "bkTgitReviewClosedEventDesc": "Code review [{review}] closed",
+    "bkTgitReviewChangeDeniedEventDesc": "Code review [{review}] change denied by {user}",
+    "bkTgitReviewChangeRequiredEventDesc": "Code review [{review}] change required by {user}",
+    "bkGithubPushEventDesc": "[{branch}] Commit [{commit}] pushed by {user}",
+    "bkGithubCreateBranchEventDesc": "Branch [{branch}] created by {user}",
+    "bkGithubCreateTagEventDesc": "Tag [{tag}] pushed by {user}",
+    "bkGithubPrEventDesc": "Pull request [{pr}] by {user} {action}",
+    "bkP4EventDesc": "Change [{change}] by {user} {action}",
+    "bkSvnCommitEventDesc": "Commit [{revision}] pushed by {user}",
+    "bkManualStartEventDesc": "Manually triggered by {user}",
+    "bkRemoteStartEventDesc": "Remote api triggered by {remoteUser}",
+    "bkServiceStartEventDesc": "Pipeline invocation triggered by {user} invoke",
+    "bkPipelineStartEventDesc": "Sub-Pipeline triggered by [{pipeline}] {user}",
+    "bkTimingStartEventDesc": "Timer triggered by {user}",
+    "bkEventReplayDesc": "Event [{event}] replayed by {user}",
+    "bkRepoEnablePacEventDesc": "Enable PAC by {user}",
+    "bkGitPushEventDesc": "[{branch}] Commit [{commit}] pushed by {user}",
+    "bkGitPushCreateEventDesc": "Branch [{branch}] created by {user}",
+    "bkGitPushDeleteEventDesc": "Branch [{branch}] deleted by {user}",
+    "bkGitIssueOpenedEventDesc": "Issue [{issue}] opened by {user}",
+    "bkGitIssueUpdatedEventDesc": "Issue [{issue}] updated by {user}",
+    "bkGitIssueClosedEventDesc": "Issue [{issue}] closed by {user}",
+    "bkGitIssueReopenedEventDesc": "Issue [{issue}] reopened by {user}",
+    "bkGitPrCreatedEventDesc": "Merge request [{mr}] opened by {user}",
+    "bkGitPrUpdatedEventDesc": "Merge request [{mr}] edited by {user}",
+    "bkGitPrClosedEventDesc": "Merge request [{mr}] closed by {user}",
+    "bkGitPrReopenedEventDesc": "Merge request [{mr}] reopened by {user}",
+    "bkGitPrPushUpdatedEventDesc": "Merge request [{mr}] push updated by {user}",
+    "bkGitPrMergedEventDesc": "Merge request [{mr}] merged by {user}",
+    "bkGitTagPushEventDesc": "[{source}] Tag [{tag}] pushed by {user}",
+    "bkGitTagDeleteEventDesc": "[{source}] Tag [{tag}] deleted by {user}",
+    "bkGitNoteEventDesc": "Note [{note}] submitted by {user}",
+    "bkGitReviewCreatedEventDesc": "Code review [{review}] created by {user}",
+    "bkGitReviewApprovedEventDesc": "Code review [{review}] approved by {user}",
+    "bkGitReviewApprovingEventDesc": "Code review [{review}] approving",
+    "bkGitReviewClosedEventDesc": "Code review [{review}] closed",
+    "bkGitReviewChangeDeniedEventDesc": "Code review [{review}] change denied by {user}",
+    "bkGitReviewChangeRequiredEventDesc": "Code review [{review}] change required by {user}",
+    "bkSubPipelineCircularDependencyErrorMessage": "Sub-pipeline [{pipeline}] called the current pipeline, forming a circular dependency",
+    "bkOtherSubPipelineCircularDependencyErrorMessage": "Sub-pipeline [{pipeline}] called [{targetPipeline}] pipeline, forming a circular dependency",
+    "bkYamlPipelineCreateSuccess": "Pipeline [{pipeline} | {version}] created successfully",
+    "bkYamlPipelineUpdateSuccess": "Pipeline [{pipeline} | {version}] updated successfully",
+    "bkYamlPipelineRenameSuccess": "Pipeline [{pipeline}|{version}] renamed successfully, {oldName} → {newName}",
+    "bkYamlPipelineDeleteVersionSuccess": "Pipeline [{pipeline} | {version}] version deleted successfully",
+    "bkYamlPipelineDeleteSuccess": "Pipeline [{name} | {version}] deleted successfully",
+    "bkYamlPipelineCreateFailed": "Failed to create pipeline",
+    "bkYamlPipelineUpdateFailed": "Pipeline [{pipeline}] update failed",
+    "bkYamlPipelineRenameFailed": "Pipeline [{pipeline}] rename failed, {oldName} → {newName}",
+    "bkYamlPipelineDependencyUpgradeFailed": "Pipeline [{pipeline}] dependency upgrade failed",
+    "bkYamlPipelineDeleteVersionFailed": "Pipeline [{pipeline} | {version}] version deletion failed",
+    "bkYamlPipelineDeleteFailed": "Pipeline [{pipeline}] deletion failed",
+    "bkYamlPipelineCloseFailed": "Pipeline [{pipeline}] failed to update instantiation status",
+    "bkYamlInstancePullRequestClosed": "MR [{mr}] has been closed, changes not published to default branch"
 }
