+++ src/frontend/locale/pipeline/en-US.json
@@ -1500,6 +1500,43 @@
   "bkTimingStartEventDesc": "Timer triggered by {user}",
   "bkEventReplayDesc": "Event [{event}] replayed by {user}",
   "bkRepoEnablePacEventDesc": "Enable PAC by {user}",
+  "bkSubPipelineCircularDependencyErrorMessage": "Sub-pipeline [{pipeline}] called the current pipeline, forming a circular dependency",
+  "bkOtherSubPipelineCircularDependencyErrorMessage": "Sub-pipeline [{pipeline}] called [{targetPipeline}] pipeline, forming a circular dependency",
+  "bkGitPushEventDesc": "[{branch}] Commit [{commit}] pushed by {user}",
+  "bkGitPushCreateEventDesc": "Branch [{branch}] created by {user}",
+  "bkGitPushDeleteEventDesc": "Branch [{branch}] deleted by {user}",
+  "bkGitIssueOpenedEventDesc": "Issue [{issue}] opened by {user}",
+  "bkGitIssueUpdatedEventDesc": "Issue [{issue}] updated by {user}",
+  "bkGitIssueClosedEventDesc": "Issue [{issue}] closed by {user}",
+  "bkGitIssueReopenedEventDesc": "Issue [{issue}] reopened by {user}",
+  "bkGitPrCreatedEventDesc": "Merge request [{mr}] opened by {user}",
+  "bkGitPrUpdatedEventDesc": "Merge request [{mr}] edited by {user}",
+  "bkGitPrClosedEventDesc": "Merge request [{mr}] closed by {user}",
+  "bkGitPrReopenedEventDesc": "Merge request [{mr}] reopened by {user}",
+  "bkGitPrPushUpdatedEventDesc": "Merge request [{mr}] push updated by {user}",
+  "bkGitPrMergedEventDesc": "Merge request [{mr}] merged by {user}",
+  "bkGitTagPushEventDesc": "[{source}] Tag [{tag}] pushed by {user}",
+  "bkGitTagDeleteEventDesc": "[{source}] Tag [{tag}] deleted by {user}",
+  "bkGitNoteEventDesc": "Note [{note}] submitted by {user}",
+  "bkGitReviewCreatedEventDesc": "Code review [{review}] created by {user}",
+  "bkGitReviewApprovedEventDesc": "Code review [{review}] approved by {user}",
+  "bkGitReviewApprovingEventDesc": "Code review [{review}] approving",
+  "bkGitReviewClosedEventDesc": "Code review [{review}] closed",
+  "bkGitReviewChangeDeniedEventDesc": "Code review [{review}] change denied by {user}",
+  "bkGitReviewChangeRequiredEventDesc": "Code review [{review}] change required by {user}",
+  "bkYamlPipelineCreateSuccess": "Pipeline [{pipeline} | {version}] created successfully",
+  "bkYamlPipelineUpdateSuccess": "Pipeline [{pipeline} | {version}] updated successfully",
+  "bkYamlPipelineRenameSuccess": "Pipeline [{pipeline}|{version}] renamed successfully, {oldName} → {newName}",
+  "bkYamlPipelineDeleteVersionSuccess": "Pipeline [{pipeline} | {version}] version deleted successfully",
+  "bkYamlPipelineDeleteSuccess": "Pipeline [{name} | {version}] deleted successfully",
+  "bkYamlPipelineCreateFailed": "Failed to create pipeline",
+  "bkYamlPipelineUpdateFailed": "Pipeline [{pipeline}] update failed",
+  "bkYamlPipelineRenameFailed": "Pipeline [{pipeline}] rename failed, {oldName} → {newName}",
+  "bkYamlPipelineDependencyUpgradeFailed": "Pipeline [{pipeline}] dependency upgrade failed",
+  "bkYamlPipelineDeleteVersionFailed": "Pipeline [{pipeline} | {version}] version deletion failed",
+  "bkYamlPipelineDeleteFailed": "Pipeline [{pipeline}] deletion failed",
+  "bkYamlPipelineCloseFailed": "Pipeline [{pipeline}] failed to update instantiation status",
+  "bkYamlInstancePullRequestClosed": "MR [{mr}] has been closed, changes not published to default branch",
   "invalidCodeMode": "Switch to Code Mode Failed",
   "draftExecRecords": "Debug Records",
   "draftExecDetail": "Debug Details",
