+++ src/frontend/locale/pipeline/zh-CN.json
@@ -1501,6 +1501,43 @@
   "bkTimingStartEventDesc": "{user} 定时触发",
   "bkEventReplayDesc": "事件 [{event}] 由 {user} 重放",
   "bkRepoEnablePacEventDesc": "{user} 开启PAC",
+  "bkSubPipelineCircularDependencyErrorMessage": "子流水线 [{pipeline}] 调用了当前流水线，形成循环依赖",
+  "bkOtherSubPipelineCircularDependencyErrorMessage": "子流水线 [{pipeline}] 调用了 [{targetPipeline}] 流水线，形成循环依赖",
+  "bkGitPushEventDesc": "[{branch}] 提交 [{commit}] 由 {user} 推送",
+  "bkGitPushCreateEventDesc": "分支 [{branch}] 由 {user} 创建",
+  "bkGitPushDeleteEventDesc": "分支 [{branch}] 由 {user} 删除",
+  "bkGitIssueOpenedEventDesc": "议题 [{issue}] 由 {user} 创建",
+  "bkGitIssueUpdatedEventDesc": "议题 [{issue}] 由 {user} 更新",
+  "bkGitIssueClosedEventDesc": "议题 [{issue}] 由 {user} 关闭",
+  "bkGitIssueReopenedEventDesc": "议题 [{issue}] 由 {user} 重新打开",
+  "bkGitPrCreatedEventDesc": "合并请求 [{mr}] 由 {user} 创建",
+  "bkGitPrUpdatedEventDesc": "合并请求 [{mr}] 由 {user} 编辑",
+  "bkGitPrClosedEventDesc": "合并请求 [{mr}] 由 {user} 关闭",
+  "bkGitPrReopenedEventDesc": "合并请求 [{mr}] 由 {user} 重新打开",
+  "bkGitPrPushUpdatedEventDesc": "合并请求 [{mr}] 由 {user} 在源分支提交代码触发",
+  "bkGitPrMergedEventDesc": "合并请求 [{mr}] 由 {user} 合并",
+  "bkGitTagPushEventDesc": "[{source}] 创建Tag [{tag}] 由 {user} 推送",
+  "bkGitTagDeleteEventDesc": "[{source}] Tag [{tag}] 由 {user} 删除",
+  "bkGitNoteEventDesc": "评论 [{note}] 由 {user} 发表",
+  "bkGitReviewCreatedEventDesc": "代码审查 [{review}] 由 {user} 创建",
+  "bkGitReviewApprovedEventDesc": "代码审查 [{review}] 由 {user} 评审通过",
+  "bkGitReviewApprovingEventDesc": "代码审查 [{review}] 评审中",
+  "bkGitReviewClosedEventDesc": "代码审查 [{review}] 已关闭",
+  "bkGitReviewChangeDeniedEventDesc": "代码审查 [{review}] 被 {user} 拒绝",
+  "bkGitReviewChangeRequiredEventDesc": "代码审查 [{review}] 由 {user} 要求修改",
+  "bkYamlPipelineCreateSuccess": "流水线 [{pipeline} | {version}] 创建成功",
+  "bkYamlPipelineUpdateSuccess": "流水线 [{pipeline} | {version}] 更新成功",
+  "bkYamlPipelineRenameSuccess": "流水线 [{pipeline}|{version}]重命名成功，{oldName} → {newName}",
+  "bkYamlPipelineDeleteVersionSuccess": "流水线 [{pipeline} | {version}] 删除版本成功",
+  "bkYamlPipelineDeleteSuccess": "流水线 [{name} | {version}] 删除成功",
+  "bkYamlPipelineCreateFailed": "创建流水线失败",
+  "bkYamlPipelineUpdateFailed": "流水线 [{pipeline}] 更新失败",
+  "bkYamlPipelineRenameFailed": "流水线 [{pipeline}]重命名失败，{oldName} → {newName}",
+  "bkYamlPipelineDependencyUpgradeFailed": "流水线 [{pipeline}] 依赖更新失败",
+  "bkYamlPipelineDeleteVersionFailed": "流水线 [{pipeline} | {version}] 版本删除失败",
+  "bkYamlPipelineDeleteFailed": "流水线 [{pipeline}] 删除失败",
+  "bkYamlPipelineCloseFailed": "流水线 [{pipeline}] 更新实例化状态失败",
+  "bkYamlInstancePullRequestClosed": "MR [{mr}]已关闭，变更未发布到默认分支",
   "invalidCodeMode": "切换到Code方式失败",
   "draftExecRecords": "调试记录",
   "draftExecDetail": "调试详情",
